coala-setup.py.jj2: Add maintainer_list
With this commit we can either provide the key maintainers
or maintainer_list to provide the list of maintainers. The former will
print everything on a single line but the latter will print everything
on different lines. If both maintainers and maintainer_list are
provided, then maintainer_list will be overridden.

Closes https://gitlab.com/coala/mobans/issues/16