coala-setup.py.jj2: Split egg name
Package names may contain a '-', which is also used
to denote a version in an egg name.
This splits the egg name on the first part which
looks like a version.
It also detects dev versions, and guesses the
prior version so it can ensure the installed
version is higher than the previously released
version.  Additionally, if the version ends in
zeros, it guesses the previous version.
This is because setuptools needs a PyPi release
for any specifiers, like `==` and `=~`, that
it encounters in the requirements.

Fixes https://gitlab.com/coala/mobans/issues/51